Wait for idle when destroying a surface. Also fixed root signature and pipeline sate object destruction in graphics device. Code renaming.

+++ Code/Engine/Source/Bruno/Core/FileStream.cpp
@@ -26,7 +26,7 @@ namespace Bruno
 
 		if (m_stream.good() && (fileAccess & FileAccess::Read) != FileAccess::None)
 		{
-			m_fileLength = std::filesystem::file_size(filename);
+			m_fileLength = static_cast<long>(std::filesystem::file_size(filename));
 		}
 	}
